requirements.txt
streamlit==1.34.0
tensorflow==2.16.1
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.3.0